pyproject.toml
[project]
name = "sites-internet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "bcrypt==4.0.1",
    "dnspython>=2.8.0",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "icalendar>=7.0.3",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.1.1",
    "pydantic[email]>=2.12.5",
    "pymongo>=4.16.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "pytz>=2026.1.post1",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]
